+++ packages/canon/.storybook/main.ts
@@ -14,8 +14,8 @@ function getAbsolutePath(value: string): any {
 const config: StorybookConfig = {
   stories: [
     '../docs/**/*.mdx',
-    '../src/components-new/**/*.mdx',
-    '../src/components-new/**/*.stories.@(js|jsx|mjs|ts|tsx)',
+    '../src/components/**/*.mdx',
+    '../src/components/**/*.stories.@(js|jsx|mjs|ts|tsx)',
   ],
   staticDirs: ['../static'],
   addons: [
