+++ packages/canon/docs/Home.mdx
@@ -1,5 +1,5 @@
 import { Unstyled } from '@storybook/blocks';
-import { Columns, Text, ComponentStatus, Banner } from './components-new';
+import { Columns, Text, ComponentStatus, Banner } from './components';
 
 <Unstyled>
