+++ packages/canon/docs/Theme.mdx
@@ -1,18 +1,15 @@
 import { Unstyled, Source } from '@storybook/blocks';
-import { Title, Text, Chip } from './components-new';
-import * as Table from './components-new/Table';
+import { Title, Text, Chip } from './components';
+import * as Table from './components/Table';
 
 <Unstyled>
 
 <Title type="h1">Theming</Title>
 <Text>
   Backstage ships with a default theme with a light and dark mode variant. The
-  themes are provided as a part of the{' '}
-  <a href="https://github.com/backstage/backstage/packages/canon">
-    @backstage/canon
-  </a>{' '}
-  package, which also includes utilities for customizing the default theme, or
-  creating completely new themes.
+  themes are provided as a part of the `@backstage/canon` package, which also
+  includes utilities for customizing the default theme, or creating completely
+  new themes.
 </Text>
 
 <Title type="h2">Light & Dark modes</Title>
