+++ packages/canon/docs/components/Text/index.tsx
