+++ packages/canon/docs/components/Text/text.css.ts
