+++ packages/canon/src/components/Box/box.mdx
@@ -1,10 +1,10 @@
 import { Canvas, Meta, Unstyled } from '@storybook/blocks';
 import * as BoxStories from './Box.stories';
-import { Chip } from '../../../docs/components-new/Chip';
-import * as Table from '../../../docs/components-new/Table';
+import { Chip } from '../../../docs/components/Chip';
+import * as Table from '../../../docs/components/Table';
 import { Box } from './Box';
 import { Stack } from '../Stack';
-import { Title, Text } from '../../../docs/components-new';
+import { Title, Text } from '../../../docs/components';
 import { Padding } from './docs/padding';
 import { listResponsiveValues, listColorValues } from '../../utils/list-values';
 import { responsiveProperties } from './sprinkles.css';
