+++ packages/canon/src/components/Box/docs/padding.tsx
