+++ packages/canon/src/components/Box/properties.ts
