+++ packages/canon/src/components/Icon/index.tsx
