+++ packages/canon/src/components/Inline/Inline.tsx
