+++ packages/canon/src/components/Inline/index.ts
