+++ packages/canon/src/utils/list-values.ts
@@ -17,7 +17,7 @@
 import {
   responsiveProperties,
   colorProperties,
-} from '../components-new/Box/sprinkles.css';
+} from '../components/Box/sprinkles.css';
 
 export const listResponsiveValues = (
   value: keyof typeof responsiveProperties.styles,
